requirements.txt
spacy
pandas
numpy
openai
python-dotenv
matplotlib
seaborn
openpyxl
tiktoken
statsmodels